Refactor cursor code
Closes: #93
