+++ src/graphql.rs
@@ -21,6 +21,27 @@ pub fn schema(database: Database) -> Schema {
         .finish()
 }
 
+fn check_paging_type(
+    after: Option<String>,
+    before: Option<String>,
+    first: Option<usize>,
+    last: Option<usize>,
+) -> anyhow::Result<PagingType> {
+    if let Some(val) = first {
+        if let Some(cursor) = after {
+            return Ok(PagingType::AfterFirst(cursor, val));
+        }
+        return Ok(PagingType::First(val));
+    }
+    if let Some(val) = last {
+        if let Some(cursor) = before {
+            return Ok(PagingType::BeforeLast(cursor, val));
+        }
+        return Ok(PagingType::Last(val));
+    }
+    Err(anyhow::anyhow!("Invalid paging type"))
+}
+
 #[cfg(test)]
 struct TestSchema {
     _dir: tempfile::TempDir, // to prevent the data directory from being deleted while the test is running
